Add console.log to observe game loop updates
Inserted a console.log inside the game loop to visualize how often the page is being redrawn.
This helps me understand the update frequency and confirms that the game loop is running
correctly at the current speed. Being able to see these updates in the browser console
makes it easier to learn how the game logic works and how the movement will behave. Also, is very cool